+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-sparkoncosmos/src/main/java/com/microsoft/azure/toolkit/intellij/sparkoncosmos/actions/SubmitClusterJobAction.java
@@ -5,7 +5,7 @@ import com.intellij.execution.RunnerAndConfigurationSettings;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.Presentation;
 import com.intellij.openapi.project.Project;
-import com.microsoft.azure.cosmos.spark.actions.CosmosSparkSelectAndSubmitAction;
+import com.microsoft.azure.hdinsight.spark.actions.CosmosSparkSelectAndSubmitAction;
 import com.microsoft.azure.hdinsight.sdk.common.azure.serverless.AzureSparkCosmosCluster;
 import com.microsoft.azure.cosmos.spark.actions.SparkAppSubmitContext;
 import com.microsoft.azure.hdinsight.spark.run.configuration.CosmosSparkConfigurationFactory;
